Add project API permissions and roles
