--- app/api/permissions.py
@@ -1,10 +1,0 @@
-from rest_framework.permissions import BasePermission
-
-
-class IsSuperuser(BasePermission):
-    """
-    Allows access only to super users.
-    """
-
-    def has_permission(self, request, view):
-        return bool(request.user and request.user.is_superuser)
